% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpenter_model.R
\name{carpenter_model}
\alias{carpenter_model}
\title{Run the simulation for Carpenter (1992)}
\usage{
carpenter_model(
  nanoplankter_diameter,
  alga_diameter,
  length_herbivore,
  death_rate_herbivore,
  p_influx_rate,
  p_outflow_rate,
  mixed_layer_depth
)
}
\arguments{
\item{nanoplankter_diameter}{diameter of nanoplankter (um) (previously diam1)}

\item{alga_diameter}{diameter of blue green alga (um) (previously diam2)}

\item{length_herbivore}{herbivore length (mm) (previously zlen)}

\item{death_rate_herbivore}{herbivore death rate (fraction per day) (previously c[3])}

\item{p_influx_rate}{P influx rate (ug P per L per day) (previously c[11])}

\item{p_outflow_rate}{P outflow rate (fraction per day) (previously outflow)}

\item{mixed_layer_depth}{mixed layer depth (m) (previously zmix)}
}
\value{
This function returns a named list, where the first entry is a
dataframe containing the results of the simulations, and the remaining
entries give the supplied arguments and intermediate results from the model.
}
\description{
This function implements the model from Carpenter's 1992
paper "Destabilization of Planktonic Ecosystems and Blooms of Blue-Green
Algae."
}
\examples{
# run the model using the mendota defaults
carpenter_model(nanoplankter_diameter = 5.0,
                alga_diameter = 40,
                length_herbivore = 1,
                death_rate_herbivore = .1,
                p_influx_rate = 0.7,
                p_outflow_rate = 0.0003,
                mixed_layer_depth = 2.5)
}
