% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpenter_pivot.R
\name{carpenter_pivot}
\alias{carpenter_pivot}
\title{Pivot Carpenter model data}
\usage{
carpenter_pivot(model_data)
}
\arguments{
\item{model_data}{A dataframe outputted by \code{carpenter_model()} or
\code{carpenter_pivot()} itself.}
}
\value{
A tibble, either with 480 rows and 3 columns or 120 rows and 5 
columns.
}
\description{
This function allows you to convert data outputted by
 \code{carpenter_model()} from "wide" to "long" format, or the other way 
 around. Depending on the desired use case for data outputted by the modeling
 function, the data may or may not already be "tidy" (and thus easier to
 use with functions from the tidyverse.) This function is just a wrapper
 around \code{\link[tidyr:pivot_wider]{pivot_wider}} and 
 \code{\link[tidyr:pivot_longer]{pivot_longer}}--see the documentation for
 these functions for more complicated use cases.
}
\examples{
# run the model using the mendota defaults
model <- carpenter_model(nanoplankter_diameter = 5.0,
                         alga_diameter = 40,
                         length_herbivore = 1,
                         death_rate_herbivore = .1,
                         p_influx_rate = 0.7,
                         p_outflow_rate = 0.0003,
                         mixed_layer_depth = 2.5)
                        
# convert the data to long format!
carpenter_pivot(model)

# running the function twice returns the original data
carpenter_pivot(carpenter_pivot(model))

}
\seealso{
\code{carpenter_model()} for the function to generate appropriate 
data and \code{carpenter_visualize()} to visualize the data.
}
