% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{carpenter_visualize}
\alias{carpenter_visualize}
\alias{carpenter_viz_phosphorus}
\alias{carpenter_viz_chlorophyll}
\alias{carpenter_viz_zooplankton}
\title{Visualize Carpenter model data}
\usage{
carpenter_visualize(model_data)

carpenter_viz_phosphorus(model_data)

carpenter_viz_chlorophyll(model_data)

carpenter_viz_zooplankton(model_data)
}
\arguments{
\item{model_data}{A dataframe outputted by \code{carpenter_model()}}
}
\value{
A ggplot object displaying the concentration of the relevant
variables over time.
}
\description{
These functions allow the user to visualize the outputs of
the \code{carpenter_model()} function. \code{carpenter_visualize()} calls
each of the other visualization functions in succession!
}
\examples{
# run the model using the mendota defaults
model <- carpenter_model(nanoplankter_diameter = 5.0,
                         alga_diameter = 40,
                         length_herbivore = 1,
                         death_rate_herbivore = .1,
                         p_influx_rate = 0.7,
                         p_outflow_rate = 0.0003,
                         mixed_layer_depth = 2.5)
                        
# plot the data using the different carpenter_vi* functions!
carpenter_viz_phosphorus(model)
carpenter_viz_chlorophyll(model)
carpenter_viz_zooplankton(model)
carpenter_visualize(model)

# you can add on layers to the output of this function
# as would any ggplot output
carpenter_viz_phosphorus(model) + ggplot2::theme_minimal()

}
\seealso{
\code{carpenter_model()} for the function to generate appropriate 
data.
}
