% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpentR.R
\docType{package}
\name{carpentR}
\alias{carpentR}
\alias{carpentR-package}
\title{Package: carpentR}
\description{
This R Package implements a model from Stephen R. Carpenter's 1992 
paper on predicting lake algal blooms using plankton dynamics and other 
variables, and was developed specifically for use in Sam Fey's Ecology 
class at Reed College.
}
\author{
\strong{Maintainer}: Simon Couch \email{simonpatrickcouch@gmail.com}

Other contributors:
\itemize{
  \item Sam Fey \email{feys@reed.edu} [conceptor]
  \item Stephen R. Carpenter [conceptor]
}

}
