% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiti_oilgas_revenue.R
\name{eiti_oilgas_revenue}
\alias{eiti_oilgas_revenue}
\title{Estimate carbon revenue for a given EITI country and year and carbon price}
\usage{
eiti_oilgas_revenue(country, year, carbonprice)
}
\arguments{
\item{country}{2 letter country code}

\item{year}{as integer}

\item{carbonprice}{carbon price per metric tonne in USD}
}
\value{
A numeric revenue estimate
}
\description{
Estimate carbon revenue for a given EITI country and year and carbon price
}
\examples{

eiti_oilgas_revenue("SN",2017,35)
}
