% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Euclidean distance}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric vector}

\item{b}{A numeric vector}
}
\value{
A numeric Euclidean distance
}
\description{
Calculate Euclidean distance
}
\examples{

x <- c(4,5,7)
y <- c(9,12,3)
euclidean(x,y)
}
