% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_revenue.R
\name{get_revenue}
\alias{get_revenue}
\title{Estimate carbon pricing revenue}
\usage{
get_revenue(carbonprice, emissions)
}
\arguments{
\item{carbonprice}{carbon price per metric tonne in USD}

\item{emissions}{carbon emissions in metric tonnes}
}
\value{
A numeric revenue estimate
}
\description{
Estimate carbon pricing revenue
}
\examples{

get_revenue(35,1000000)
}
