% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{get_wb_indicators}
\alias{get_wb_indicators}
\title{Get world bank indicators}
\usage{
get_wb_indicators(world_bank_raw, year, initial_indicator_set)
}
\arguments{
\item{world_bank_raw}{Raw world bank development indicator data}

\item{year}{The year to get the data from}

\item{indicator_set}{a vector of indicator codes to filter}
}
\value{
A dataframe of cleaned indicators
}
\description{
Get world bank indicators, and transform values in USD per capita and in %GDP to totals in USD
}
\examples{

initial_indicator_set <- c(
'GC.TAX.TOTL.GD.ZS',
'MS.MIL.XPND.CD',
'NY.ADJ.AEDU.CD',
'SH.XPD.CHEX.PC.CD',
'ST.INT.XPND.CD',
'NE.CON.GOVT.CD'
)

get_wb_indicators(world_bank_raw,2017,initial_indicator_set)


}
