% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilgas_to_co2.R
\name{oilgas_to_co2}
\alias{oilgas_to_co2}
\title{Convert oil and gas production to CO2 emissions}
\usage{
oilgas_to_co2(production_data)
}
\arguments{
\item{production_data}{generated from read_eiti_production}
}
\value{
A numeric estimate of CO2 emissions
}
\description{
Convert oil and gas production to CO2 emissions
}
\examples{


sn17 <- read_eiti_production("SN",2017)
oilgas_to_co2(sn17)
}
