% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plot indicator}
\usage{
plot_indicator(indicator_data, revenue, revenue_year)
}
\arguments{
\item{indicator_data}{a time series of the indicator data, with year and value as column names. Value should be in USD}

\item{revenue}{revenue in USD}

\item{revenue_year}{the year that the revenue data is calculated for}
}
\value{
A ggplot object
}
\description{
Plot indicator
}
\examples{


}
