% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eiti.R
\name{read_eiti}
\alias{read_eiti}
\title{Read in EITI production data}
\usage{
read_eiti(country, year, type)
}
\arguments{
\item{country}{2 letter country code}

\item{year}{as integer}

\item{type}{must be 'production' or 'revenue'}
}
\value{
A data frame with one row for oil production/revenue and one for gas production/revenue
}
\description{
Read EITI production or revenue data for a given country and year
}
\examples{

read_eiti_production("SN",2017)
}
