% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eiti_production.R
\name{read_eiti_production}
\alias{read_eiti_production}
\title{Read in EITI production data}
\usage{
read_eiti_production(country, year)
}
\arguments{
\item{country}{2 letter country code}

\item{year}{as integer}
}
\value{
A data frame with one row for oil production and one for gas production
}
\description{
Read EITI production data for a given country and year
}
\examples{

read_eiti_production("SN",2017)
}
