% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{update_indicator_question_template}
\alias{update_indicator_question_template}
\title{Update indicator-question template}
\usage{
update_indicator_question_template(dir = "data-raw")
}
\arguments{
\item{dir}{directory containing the files indicators.csv and questions.csv}
}
\value{
A csv file 'indicator_question_template.csv' written to data-raw
}
\description{
Write a blank template of expected scores for a given set of indicators and questions
}
\examples{


update_indicator_question_template()

}
