% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_carpenter}
\alias{visualize_carpenter}
\alias{visualize_phosphorus}
\alias{visualize_chlorophyll}
\alias{visualize_zooplankton}
\title{Visualize Carpenter model data}
\usage{
visualize_carpenter(model_data)

visualize_phosphorus(model_data)

visualize_chlorophyll(model_data)

visualize_zooplankton(model_data)
}
\arguments{
\item{model_data}{A dataframe outputted by \code{carpenter_model()}}
}
\value{
A ggplot object displaying the concentration of the relevant
variables over time.
}
\description{
These functions allow the user to visualize the outputs of
the \code{carpenter_model()} function. \code{visualize_carpenter()} calls
each of the other visualization functions in succession!
}
\examples{
# run the model using the mendota defaults
model <- carpenter_model(nanoplankter_diameter = 5.0,
                         alga_diameter = 40,
                         length_herbivore = 1,
                         death_rate_herbivore = .1,
                         p_influx_rate = 0.7,
                         p_outflow_rate = 0.0003,
                         mixed_layer_depth = 2.5)
                        
# plot the data using the different visualize_* functions!
visualize_phosphorus(model)
visualize_chlorophyll(model)
visualize_zooplankton(model)
visualize_carpenter(model)

}
\seealso{
\code{carpenter_model()} for the function to generate appropriate 
data.
}
