% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpenter_model.R
\name{carpenter_model}
\alias{carpenter_model}
\title{Run the simulation for Carpenter (1992)}
\usage{
carpenter_model(
  nanoplankter_diameter,
  alga_diameter,
  length_herbivore,
  death_rate_herbivore,
  p_influx_rate,
  p_outflow_rate,
  mixed_layer_depth
)
}
\arguments{
\item{nanoplankter_diameter}{diameter of nanoplankter (micrometers):
acceptable range is [2, 5]}

\item{alga_diameter}{diameter of blue green alga (micrometers): acceptable
range is [50, 2000]}

\item{length_herbivore}{herbivore length (millimeters): acceptable range
is [.3, 2.5]}

\item{death_rate_herbivore}{herbivore death rate (fraction per day):
acceptable range is [.05, .3]}

\item{p_influx_rate}{P influx rate (ug P per L per day): acceptable range
is [.1, .4]}

\item{p_outflow_rate}{P outflow rate (fraction per day): acceptable range
is [0, 1]}

\item{mixed_layer_depth}{mixed layer depth (meters): acceptable range is
[3.7, 10.2]}
}
\value{
This function returns a tibble, with columns:
\describe{
\item{day}{the simulation day}
\item{phosphorus}{phosphorus concentration (micrograms per liter)}
\item{algal_chlorophyll}{algal chlorophyll concentration (micrograms 
per liter)}
\item{blue_green_chlorophyll}{blue green chlorophyll concentration 
(micrograms per liter)}
\item{zooplankton_biomass}{zooplankton biomass (micrograms per liter)}
}
}
\description{
This function implements the model from Carpenter's 1992
paper "Destabilization of Planktonic Ecosystems and Blooms of Blue-Green
Algae."
}
\examples{
# run the model using the mendota defaults
carpenter_model(nanoplankter_diameter = 5.0,
                alga_diameter = 40,
                length_herbivore = 1,
                death_rate_herbivore = .1,
                p_influx_rate = 0.7,
                p_outflow_rate = 0.0003,
                mixed_layer_depth = 2.5)
}
