% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb_indicators.R
\name{get_wb_indicators}
\alias{get_wb_indicators}
\title{Get world bank indicators}
\usage{
get_wb_indicators(years, data_dir = "data-raw")
}
\arguments{
\item{years}{A numeric vector of the years to get the data from}

\item{data_dir}{directory containing raw world bank and indicator data. Named world_bank_raw.csv and indicators.csv, respectively}
}
\value{
A dataframe of cleaned indicators
}
\description{
Get world bank indicators, and transform values in USD per capita and in %GDP to totals in USD
}
\examples{


get_wb_indicators(2017)


}
